Django Project/requirements.txt
django ==  4.2.7
djangorestframework == 3.12.4
Pillow == 10.1.0
